==> 3_YearTrend.Rmd <==
Extract Year information
```{r}
meta$year <- unlist(lapply(as.character(meta$release_date),function(x) if(grepl("-", x)){substr(x, 1,4)}else{substr(x, nchar(x)-3, nchar(x))}))
```

drawing the number of movies each year
```{r fig.height=2, fig.width=10}
df <- data.frame(table(meta$year))
colnames(df) <- c("Year","Number")
df <- df[-1, ]
df$Year <- as.numeric(as.character(df$Year))
ggplot(df, aes(Year,Number)) + geom_bar(stat = "identity")
#plot(table(meta$year))
```

Divide the year into six time periods
```{r}
yrs <- c("1950","1970","1990","2000","2010","2020")
meta$peri <- unlist(lapply(meta$year, function(x) yrs[sum(x>=yrs)+1]))
```

```{r}
keyw.meta <- meta[rownames(keyw.mat.trimmed),]
```

```{r}
library("RColorBrewer")
```


```{r fig.height=6, fig.width=6}
cols <- rev(brewer.pal(n = 6, name = "RdBu"))
plot(res.mca$ind$coord[,c(1,2)], col=cols[as.factor(keyw.meta$peri)])
plot(res.mca$ind$coord[,c(3,4)], col=cols[as.factor(keyw.meta$peri)])
```

```{r}
cols <- rev(brewer.pal(n = 6, name = "RdBu"))
plot3d(res.mca$ind$coord[,c(1:3)], col=cols[as.factor(keyw.meta$peri)])
```


```{r fig.width=5, fig.height=6}
cols <- rev(brewer.pal(n = 6, name = "RdBu"))
plot_data = plot_ly(x = res.mca$ind$coord[,1], y =res.mca$ind$coord[,2], 
            z = res.mca$ind$coord[,3], type="scatter3d", mode="markers", size = 1, color=as.factor(keyw.meta$peri), colors = cols)
plot_data
```

```{r}
library(reshape2)
```


```{r}
df <- table(keyw.meta$peri, bin.clus)
df <- df/apply(df,1,sum)
df <- melt(df)
colnames(df) <- c("Year","Cluster", "Fraction")
df$Cluster <- factor(df$Cluster)
ggplot(df, aes(x=Year, y=Fraction, fill=Cluster)) + 
    geom_area(alpha=0.6 , size=1, colour="black")
```
